issue-194 비밀번호 UI 개선 - 다른 앱을 참고하여 비밀번호 UI 개선
