+++ app/src/main/AndroidManifest.xml
@@ -20,8 +20,10 @@
         <activity
             android:name=".view.setting.SettingActivity"
             android:label="@string/title_activity_setting"
-            android:theme="@style/AppTheme.NoActionBar"></activity>
-        <activity android:name=".view.password.PasswordSelectActivity" />
+            android:theme="@style/AppTheme.NoActionBar" />
+        <activity
+            android:name=".view.password.LegacyPasswordSelectActivity"
+            android:theme="@style/AppTheme.NoActionBar" />
         <activity
             android:name=".view.password.PasswordActivity"
             android:windowSoftInputMode="stateHidden" />
