+++ app/src/main/res/anim/anim_left_to_right.xml
@@ -3,7 +3,7 @@
     android:shareInterpolator="false">
 
     <translate
-        android:duration="700"
+        android:duration="300"
         android:fromXDelta="0%"
         android:toXDelta="0%"
         android:fromYDelta="0%"
