+++ app/src/main/res/anim/anim_right_to_left.xml
@@ -3,7 +3,7 @@
     android:shareInterpolator="false">
 
     <translate
-        android:duration="700"
+        android:duration="300"
         android:fromXDelta="0%"
         android:toXDelta="100%"
         android:fromYDelta="0%"
