+++ app/src/main/res/drawable/button_pressed.xml
@@ -4,8 +4,8 @@
 
     <stroke
         android:width="1dip"
-        android:color="@color/light_blue" />
+        android:color="@android:color/transparent" />
 
-    <solid android:color="@color/light_blue" />
+    <solid android:color="@color/gray" />
 
 </shape>
