+++ app/src/main/res/drawable/edittext_normal.xml
@@ -1,7 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle" >
+    android:shape="line" >
 
-    <solid android:color="#40DDDDDD" />
+
+    <!--형태 모형을 단색으로 채움.-->
+    <solid android:color="@color/white" />
+
+    <!--테두리-->
+    <stroke
+        android:dashWidth="20dp"
+        android:width="2dip"
+        android:color="#FFDDDDDD" />
 
 </shape>
