+++ app/src/main/res/drawable/edittext_pressed.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle" >
+    android:shape="line" >
 
-    <solid android:color="#40DDDDDD" />
+    <solid android:color="@color/white" />
 
     <stroke
         android:width="2dip"
