+++ app/src/main/res/layout/activity_password.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:tools="http://schemas.android.com/tools"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <data>
 
@@ -10,42 +11,44 @@
 
     </data>
 
-    <ScrollView
-        android:layout_width="match_parent"
+    <RelativeLayout
+        android:id="@+id/ll_password"
+        android:layout_width="wrap_content"
         android:layout_height="match_parent"
-        android:layout_gravity="center">
+        android:background="@color/white"
+        tools:context=".view.password.PasswordActivity">
 
-        <LinearLayout
-            android:id="@+id/ll_password"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:gravity="center"
-            android:orientation="vertical"
-            tools:context=".view.password.PasswordActivity">
+
+        <RelativeLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent">
 
             <TextView
                 android:id="@+id/tv_message"
-                android:layout_width="fill_parent"
+                android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:gravity="center"
                 android:paddingTop="16dp"
+                android:layout_marginTop="150dp"
                 android:paddingBottom="16dp"
                 android:singleLine="true"
                 android:textColor="#FF464646"
-                android:textSize="18sp" />
+                android:hint="비밀번호를 입력해주세요."
+                android:textSize="16dp" />
+
 
             <LinearLayout
-                android:id="@+id/ll_passcodes"
+                android:id="@+id/ll_passwords"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
+                android:layout_below="@+id/tv_message"
+                android:gravity="center_horizontal"
                 android:orientation="horizontal">
 
                 <EditText
                     android:id="@+id/et_password_one"
                     style="@style/PasswordInputTheme">
 
-                    <requestFocus />
                 </EditText>
 
                 <EditText
@@ -65,7 +68,8 @@
             <TableLayout
                 android:id="@+id/tl_keys"
                 android:layout_width="match_parent"
-                android:layout_height="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_alignParentBottom="true"
                 android:paddingTop="30dp"
                 android:shrinkColumns="*"
                 android:stretchColumns="*">
@@ -74,19 +78,19 @@
 
                     <Button
                         android:id="@+id/button_password_one"
-                        style="@style/PasswordButtonTheme"
+                        style="@style/PasswordNumberButtonTheme"
                         android:onClick="@{v->activity.onNumberButtonClick(v.getId())}"
                         android:text="1" />
 
                     <Button
                         android:id="@+id/button_password_two"
-                        style="@style/PasswordButtonTheme"
+                        style="@style/PasswordNumberButtonTheme"
                         android:onClick="@{v->activity.onNumberButtonClick(v.getId())}"
                         android:text="2" />
 
                     <Button
                         android:id="@+id/button_password_three"
-                        style="@style/PasswordButtonTheme"
+                        style="@style/PasswordNumberButtonTheme"
                         android:onClick="@{v->activity.onNumberButtonClick(v.getId())}"
                         android:text="3" />
                 </TableRow>
@@ -95,19 +99,19 @@
 
                     <Button
                         android:id="@+id/button_password_four"
-                        style="@style/PasswordButtonTheme"
+                        style="@style/PasswordNumberButtonTheme"
                         android:onClick="@{v->activity.onNumberButtonClick(v.getId())}"
                         android:text="4" />
 
                     <Button
                         android:id="@+id/button_password_five"
-                        style="@style/PasswordButtonTheme"
+                        style="@style/PasswordNumberButtonTheme"
                         android:onClick="@{v->activity.onNumberButtonClick(v.getId())}"
                         android:text="5" />
 
                     <Button
                         android:id="@+id/button_password_six"
-                        style="@style/PasswordButtonTheme"
+                        style="@style/PasswordNumberButtonTheme"
                         android:onClick="@{v->activity.onNumberButtonClick(v.getId())}"
                         android:text="6" />
                 </TableRow>
@@ -116,19 +120,19 @@
 
                     <Button
                         android:id="@+id/button_password_seven"
-                        style="@style/PasswordButtonTheme"
+                        style="@style/PasswordNumberButtonTheme"
                         android:onClick="@{v->activity.onNumberButtonClick(v.getId())}"
                         android:text="7" />
 
                     <Button
                         android:id="@+id/button_password_eight"
-                        style="@style/PasswordButtonTheme"
+                        style="@style/PasswordNumberButtonTheme"
                         android:onClick="@{v->activity.onNumberButtonClick(v.getId())}"
                         android:text="8" />
 
                     <Button
                         android:id="@+id/button_password_nine"
-                        style="@style/PasswordButtonTheme"
+                        style="@style/PasswordNumberButtonTheme"
                         android:onClick="@{v->activity.onNumberButtonClick(v.getId())}"
                         android:text="9" />
                 </TableRow>
@@ -137,7 +141,7 @@
 
                     <Button
                         android:id="@+id/button_password_clear"
-                        style="@style/PasswordButtonTheme"
+                        style="@style/PasswordNumberButtonTheme"
                         android:clickable="false"
                         android:focusable="false"
                         android:focusableInTouchMode="false"
@@ -146,21 +150,22 @@
 
                     <Button
                         android:id="@+id/button_password_zero"
-                        style="@style/PasswordButtonTheme"
+                        style="@style/PasswordNumberButtonTheme"
                         android:onClick="@{v->activity.onNumberButtonClick(v.getId())}"
                         android:text="0" />
 
                     <Button
                         android:id="@+id/button_password_erase"
-                        style="@style/PasswordButtonTheme"
+                        style="@style/PasswordNumberButtonTheme"
                         android:onClick="@{v-> activity.onDeleteNumberButton(v.getId())}"
                         android:text="Del" />
                 </TableRow>
 
             </TableLayout>
+        </RelativeLayout>
 
-        </LinearLayout>
 
-    </ScrollView>
+
+    </RelativeLayout>
 
 </layout>
