--- app/src/main/res/layout/activity_password_select.xml
@@ -1,34 +1,0 @@
-<layout xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <data>
-
-        <variable
-            name="activity"
-            type="teamh.boostcamp.myapplication.view.password.PasswordSelectActivity" />
-    </data>
-
-    <LinearLayout
-        android:id="@+id/ll_password_select"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical">
-
-        <Button
-            android:id="@+id/tv_password_set_button"
-            android:layout_width="match_parent"
-            android:layout_height="56dp"
-            android:gravity="center"
-            android:onClick="@{v->activity.onButtonClick(v.getId())}" />
-
-        <Button
-            android:id="@+id/tv_password_change_button"
-            android:layout_width="match_parent"
-            android:layout_height="56dp"
-            android:gravity="center"
-            android:onClick="@{v->activity.onButtonClick(v.getId())}"
-            android:text="@string/password_change_text" />
-
-    </LinearLayout>
-</layout>
-
-
