+++ app/src/main/res/layout/activity_setting.xml
@@ -21,7 +21,7 @@
             android:theme="@style/AppTheme.AppBarOverlay">
 
             <androidx.appcompat.widget.Toolbar
-                android:id="@+id/toolbar"
+                android:id="@+id/toolbar_setting"
                 android:layout_width="match_parent"
                 android:layout_height="?attr/actionBarSize"
                 android:background="?attr/colorPrimary"
@@ -33,7 +33,8 @@
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_gravity="center"
-                    android:text="@string/setting_title" />
+                    android:text="@string/setting_title"
+                    android:textStyle="bold" />
 
             </androidx.appcompat.widget.Toolbar>
